Complete rewrite of module alias resolving
Complete rewrite of the module alias resolving to support new features
required for a project.
Support for a new flag was added IGNORE_ABSOLUTE. This flag will allow
disable of absolute path enforcement in context that can only determine
this at runtime - this precedes the root configuration key.
React path fixing with support for specialize files is hidden
behind a import path prefix from now on: "auto:".
The new prefix is trimmed and ignored if the configuration key react is
false or not defined, this allows code reuse between configuration file
changes.